requirements.txt
torch>=1.12
torchvision>=0.13
opencv-python>=4.7.0.72
Pillow>=9.1.0
numpy>=1.22
matplotlib>=3.5
tqdm>=4.64
